# api/openapi.yaml
openapi: '3.0.0'
info:
  version: '1.0.0'
  title: 'OpenSchool API'
  description: Interact with your childrens' teachers and their classes.

servers:
  - url: http://localhost:8080

security:
  - bearerAuth: []

paths:
  /v1/auth/me:
    get:
      operationId: authCurrentUser
      summary: Use a JWT to get the currently-authenticated user.
      tags: [auth]
      responses:
        200:
          description: A JWT to be used for authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/auth/login:
    post:
      operationId: authLogin
      summary: Generate a JWT to use as a bearer token for authentication.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        200:
          description: A JWT to be used for authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/classes:
    get:
      operationId: classesList
      summary: List all classes
      tags: [classes]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: perPage
          schema:
            type: integer
          description: The number of results to retrieve in each page.
        - in: query
          name: page
          schema:
            type: integer
          description: The page to load.
      responses:
        '200':
          description: A list of classes and pagination details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassesListResponse'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: classesCreate
      summary: Create a new class
      tags: [classes]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassesCreateRequest'
      responses:
        '201':
          description: The created class, populated with metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassesCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/classes/{id}:
    get:
      operationId: classesGet
      summary: Get a class by its CUID
      tags: [classes]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Cuid'
          required: true
      responses:
        '200':
          description: The class found for the CUID provided.
          content:
            application/json:
              schema:
                type: object
                required:
                  - class
                properties:
                  class:
                    $ref: '#/components/schemas/Class'
        404:
          description: No class was found with that ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: classesUpdate
      summary: Update a class by its CUID
      tags: [classes]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Cuid'
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassesUpdateRequest'
      responses:
        '200':
          description: The class found for the CUID provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassesUpdateResponse'
        400:
          description: There was a problem with your input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: No class was found with that ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: classesDelete
      summary: Delete a class by its CUID
      tags: [classes]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Cuid'
          required: true
      responses:
        '200':
          description: The class found for the CUID provided.
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                properties:
                  ok:
                    type: boolean
                    example: true
        404:
          description: No class was found with that ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/teachers:
    get:
      operationId: teachersList
      summary: List all teachers
      tags: [teachers]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: perPage
          schema:
            type: integer
          description: The number of results to retrieve in each page.
        - in: query
          name: page
          schema:
            type: integer
          description: The page to load.
      responses:
        '200':
          description: A list of classes and pagination details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeachersListResponse'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: teachersCreate
      summary: Create a new teacher
      tags: [teachers]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeachersCreateRequest'
      responses:
        '201':
          description: The created teacher, populated with metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeachersCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/teachers/{id}:
    get:
      operationId: teachersGet
      summary: Get a teacher by its CUID
      tags: [teachers]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Cuid'
          required: true
      responses:
        '200':
          description: The teacher found for the CUID provided.
          content:
            application/json:
              schema:
                type: object
                required:
                  - class
                properties:
                  teacher:
                    $ref: '#/components/schemas/Teacher'
        404:
          description: No teacher was found with that ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: teachersUpdate
      summary: Update a teacher by its CUID
      tags: [teachers]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Cuid'
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeachersUpdateRequest'
      responses:
        '200':
          description: The teacher found for the CUID provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeachersUpdateResponse'
        400:
          description: There was a problem with your input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: No teacher was found with that ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: teachersDelete
      summary: Delete a teacher by its CUID
      tags: [teachers]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Cuid'
          required: true
      responses:
        '200':
          description: The teacher found for the CUID provided.
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                properties:
                  ok:
                    type: boolean
                    example: true
        404:
          description: No teacher was found with that ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/students:
    get:
      operationId: studentsList
      summary: List all students
      tags: [students]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: perPage
          schema:
            type: integer
          description: The number of results to retrieve in each page.
        - in: query
          name: page
          schema:
            type: integer
          description: The page to load.
      responses:
        '200':
          description: A list of students and pagination details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentsListResponse'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: studentsCreate
      summary: Create a new student
      tags: [students]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentsCreateRequest'
      responses:
        '201':
          description: The created student, populated with metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentsCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/students/{id}:
    get:
      operationId: studentsGet
      summary: Get a student by its CUID
      tags: [students]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Cuid'
          required: true
      responses:
        '200':
          description: The student found for the CUID provided.
          content:
            application/json:
              schema:
                type: object
                required:
                  - student
                properties:
                  teacher:
                    $ref: '#/components/schemas/Student'
        404:
          description: No student was found with that ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: studentsUpdate
      summary: Update a student by its CUID
      tags: [students]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Cuid'
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentsUpdateRequest'
      responses:
        '200':
          description: The student found for the CUID provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentsUpdateResponse'
        400:
          description: There was a problem with your input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: No teacher was found with that ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: studentsDelete
      summary: Delete a student by its CUID
      tags: [students]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Cuid'
          required: true
      responses:
        '200':
          description: The student found for the CUID provided.
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                properties:
                  ok:
                    type: boolean
                    example: true
        404:
          description: No student was found with that ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    PaginationData:
      type: object
      required:
        - page
        - perPage
        - total
        - nextUrl
        - prevUrl
        - firstUrl
        - lastUrl
      properties:
        page:
          type: integer
          minimum: 1
          example: 2
        perPage:
          type: integer
          minimum: 1
          maximum: 100
          example: 25
        total:
          type: integer
          minimum: 0
          example: 75
        nextUrl:
          type: string
          example: /v1/classes?page=3
        prevUrl:
          type: string
          example: /v1/classes?page=1
        firstUrl:
          type: string
          example: /v1/classes?page=1
        lastUrl:
          type: string
          example: /v1/classes?page=3

    Cuid:
      type: string
      description: A cuid
      example: cjld2cjxh0000qzrmn831i7rn

    DateTime:
      type: string
      description: An RFC3339 date/time string
      example: '1985-04-12T23:20:50.52Z'

    AuthLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string

    AuthLoginResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string

    Class:
      type: object
      required:
        - id
        - name
        - displayName
        - description
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/Cuid'
        name:
          type: string
          example: mathsOneOhOne
        displayName:
          type: string
          example: Maths 101
        description:
          type: string
          example: Basic maths class
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'

    ClassList:
      description: An array of Classes
      type: array
      items:
        $ref: '#/components/schemas/Class'

    ClassesListResponse:
      description: The response for the /v1/classes endpoint
      type: object
      required:
        - pagination
        - classes
      properties:
        pagination:
          $ref: '#/components/schemas/PaginationData'
        classes:
          $ref: '#/components/schemas/ClassList'

    ClassesCreateRequest:
      type: object
      required:
        - displayName
      properties:
        name:
          type: string
          description: The name of the Class
          example: mathsOneOhOne
        displayName:
          type: string
          description: The display name of the Class
          example: Maths 101
        description:
          type: string
          description: The description of the Class
          example: Basic maths class

    ClassesCreateResponse:
      type: object
      required:
        - class
      properties:
        class:
          $ref: '#/components/schemas/Class'

    ClassesUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: The name of the Class
          example: mathsOneOhOne
        displayName:
          type: string
          description: The display name of the Class
          example: Maths 101
        description:
          type: string
          description: The description of the Class
          example: Basic maths class

    ClassesUpdateResponse:
      type: object
      required:
        - class
      properties:
        class:
          $ref: '#/components/schemas/Class'

    Student:
      type: object
      required:
        - id
        - fullName
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/Cuid'
        fullName:
          type: string
          example: John Doe
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'

    StudentList:
      description: An array of Students
      type: array
      items:
        $ref: '#/components/schemas/Student'

    StudentsListResponse:
      description: The response for the /v1/teachers endpoint
      type: object
      required:
        - pagination
        - students
      properties:
        pagination:
          $ref: '#/components/schemas/PaginationData'
        students:
          $ref: '#/components/schemas/StudentList'

    StudentsCreateRequest:
      type: object
      required:
        - fullName
      properties:
        fullName:
          type: string
          example: John Doe

    StudentsCreateResponse:
      type: object
      required:
        - student
      properties:
        student:
          $ref: '#/components/schemas/Student'

    StudentsUpdateRequest:
      type: object
      required:
        - fullName
      properties:
        fullName:
          type: string
          example: John Doe

    StudentsUpdateResponse:
      type: object
      required:
        - student
      properties:
        student:
          $ref: '#/components/schemas/Student'

    Teacher:
      type: object
      required:
        - id
        - fullName
        - email
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/Cuid'
        fullName:
          type: string
          example: John Doe
        email:
          type: string
          example: john.doe@myschool.edu
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'

    TeacherList:
      description: An array of Teachers
      type: array
      items:
        $ref: '#/components/schemas/Teacher'

    TeachersListResponse:
      description: The response for the /v1/teachers endpoint
      type: object
      required:
        - pagination
        - teachers
      properties:
        pagination:
          $ref: '#/components/schemas/PaginationData'
        teachers:
          $ref: '#/components/schemas/TeacherList'

    TeachersCreateRequest:
      type: object
      required:
        - fullName
        - email
      properties:
        fullName:
          type: string
          example: John Doe
        email:
          type: string
          example: john.doe@myschool.edu

    TeachersCreateResponse:
      type: object
      required:
        - teacher
      properties:
        teacher:
          $ref: '#/components/schemas/Teacher'

    TeachersUpdateRequest:
      type: object
      required:
        - fullName
        - email
      properties:
        fullName:
          type: string
          example: John Doe
        email:
          type: string
          example: john.doe@school.edu

    TeachersUpdateResponse:
      type: object
      required:
        - teacher
      properties:
        teacher:
          $ref: '#/components/schemas/Teacher'

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          description: The HTTP error code
          type: integer
          example: 404
        message:
          description: A human readable error message
          type: string
          example: 'Not found'

    MultiError:
      type: object
      required:
        - code
        - messages
      properties:
        code:
          description: The HTTP error code
          type: integer
          example: 404
          minimum: 400
          maximum: 599
        messages:
          type: array
          items:
            type: string
            description: A human readable error message